moved classes to models.py
